Refactoring all the stuff...
