+++ nls/root/strings.js
@@ -6,6 +6,7 @@ define({
     BUTTON_OK:                          "OK",
     BUTTON_RESTORE_DEFAULTS:            "Restore defaults",
     BUTTON_SAVE:                        "Save",
+    DEBUG_MODE_SETTING:                 "DEBUG mode &mdash; Leave this OFF unless you need to find a problem with the extension.",
     INSERT_SNIPPET:                     "Insert snippet",
     RESTART:                            "Restart",
     Q_RESTART_BRACKETS:                 "Do you wish to restart Brackets to apply new settings?",
