+++ src/Events.js
@@ -1,9 +1,9 @@
 define(function (require, exports) {
     "use strict";
 
-    /**
-     * List of Events to be used in the extension.
-     */
-    exports.SNIPPETS_LOADED = "snippets.loaded";
+    // List of Events to be used in the extension
+    exports.EXTENSION_INIT  = "EXTENSION_INIT";
+    exports.SNIPPETS_LOADED = "SNIPPETS_LOADED";
+    exports.TRIGGER_SNIPPET = "TRIGGER_SNIPPET";
 
 });
