+++ templates/settings-dialog.html
@@ -25,6 +25,17 @@
             <input class="span8" type="text" settingsProperty="showSnippetsPanelShortcut" />
         </div>
         <!-- /shortcuts -->
+        <hr>
+        <!-- debug -->
+        <div class="row-fluid">
+            <div class="span12">
+                <label>
+                    <input type="checkbox" settingsProperty="debugMode" />
+                    {{{DEBUG_MODE_SETTING}}}
+                </label>
+            </div>
+        </div>
+        <!-- /debug -->
 
     </div>
     <div class="modal-footer">
